backend/requirements.txt
fastapi
uvicorn
python-multipart
pyshark
pymongo
python-dotenv
